Add captions
I can't make it hang over the right side of the imagery column like
in the designs because of the mix blend mode. If you do a relative
position and a negative right position to get the offset over the
column, this creates a new stacking context which prevents the blend
from working on the text outside the column container. And since the
text colour is white, you can see the bit that overhangs on the white
background.

Other things I've tried but can't get to work:
 - Use position sticky with a negative right position or negative right
   margin (can't get the text to overhang like with relative)
 - Hardcode last 20px of text to black (not sure how to make it break
   across characters nicely)
 - Changing the grid to let the caption sit outside the imagery column
   (breaks everything else)