Implement autoscroll
This smooth scrolling library implements the scroll momentum and makes
clicking the anchor nav elements feel way better. All works very nicely.
https://github.com/monsieurdd/mrD-SmoothScroller

The only nice to have would be some momentum on the click and drag
release of the imagery elements. Which is probably doable with a
requestAnimationFrame kinda thing and the right decelleration function.