Project page mobile styling
This required a refactor to the markup to remove overflow from the
imagery cards themselves which allows the captions to be put inside each
card. The overflow now goes on a sub-container to clip the content, and
the captions can sit more naturaly within their container which works
better on mobile.

There's probably a load of crap I can clear out from this CSS, but it's
working