Implement infinite looping scroll
Works on mobile and desktop by creating clones of first and last
articles at either end of the page, and instant scrolling to their
original counterparts when they cross the relevant edge of the viewport.

I had to remove that little smooth scrolling library we were using
because it was preventing our jumps from working (probably because it's
doing fancy animationFrame stuff jumping perfect numbers of pixels each
frame to fake the momentum). But we can probably find another solution
for that.